ws: move errorcodes into public module
