+++ lib/internal/websockets/Opcodes.js
@@ -1,5 +1,5 @@
 'use strict'
-const WSErrorCodes = require('internal/websockets/WSErrorCodes');
+const WSErrorCodes = require('websockets').WSErrorCodes;
 
 /**
  * utils
