+++ lib/internal/websockets/Sender.js
@@ -7,7 +7,7 @@
 const events = require('events');
 const util = require('util');
 const EventEmitter = events.EventEmitter;
-const WSErrorCodes = require('internal/websockets/WSErrorCodes');
+const WSErrorCodes = require('websockets').WSErrorCodes;
 const BufferUtil = require('internal/websockets/BufferUtil');
 const PerMessageDeflate = require('internal/websockets/PerMessageDeflate');
