+++ lib/websockets.js
@@ -22,15 +22,28 @@ const EventEmitter = require('events').EventEmitter;
 
 // Default protocol version
 const protocolVersion = 13;
-
 // Close timeout
 const closeTimeout = 30 * 1000; // Allow 30 seconds to terminate the connection cleanly
+var has = Object.prototype.hasOwnProperty;
 
-/**
- * utils
- */
-
- var has = Object.prototype.hasOwnProperty;
+const WSWSErrorCodes = {
+  1000: 'normal',
+  1001: 'going away',
+  1002: 'protocol error',
+  1003: 'unsupported data',
+  1004: 'reserved',
+  1005: 'reserved for extensions',
+  1006: 'reserved for extensions',
+  1007: 'inconsistent or invalid data',
+  1008: 'policy violation',
+  1009: 'message too big',
+  1010: 'extension handshake missing',
+  1011: 'an unexpected condition prevented the request from being fulfilled',
+  isValidErrorCode: function(code) {
+    return (code >= 1000 && code <= 1011 && code != 1004 && code != 1005 && code != 1006) ||
+         (code >= 3000 && code <= 4999);
+  }
+};
 
  /**
   * An auto incrementing id which we can use to create "unique" Ultron instances
@@ -1120,4 +1133,5 @@ function cleanupWebsocketResources(error) {
 module.exports = WebSocket;
 // TODO(eljefedelrodeodeljefe): do we need to expose this via module?
 module.exports.Server = require('internal/websockets/WebSocketServer');
+module.exports.WSWSErrorCodes = WSWSErrorCodes
 module.exports.buildHostHeader = buildHostHeader
